+++ index.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8" />
   <link rel="icon" href="/favicon.ico" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Ubarber</title>
+  <title>Ubarber Places</title>
 </head>
 
 <body>
