+++ src/views/Login.vue
@@ -1,38 +1,118 @@
 <template>
-<div id="cont-area" class="w-100 h-100 d-flex justify-content-center align-items-center">
-    <div id="sub-cont-area" class="p-4">
-        <form @submit.prevent="login()">
-            <div id="logo"><img  src="../assets/imgs/logo-black.png" alt="50px"></div>
-            <h2 class="mb-4">Login como gerenciador</h2>
-            <div class="ipt-sub-area mb-4">
-                <div class="title-ipt">E-mail</div>
-                <input placeholder="E-mail" v-bind:class="{'is-invalid': error.fun_email}" class="ipt form-control form-control-lg" type="email" name="fun_email" v-model="user.fun_email" >
-                <div class="invalid-feedback color-danger" id="fun_email" >{{error.fun_email}}</div>
-            </div>
-            <div class="ipt-sub-area mb-4">
-                <div class="title-ipt">Senha</div>
-                <input placeholder="Senha" v-bind:class="{'is-invalid': error.fun_senha}"  class="ipt form-control form-control-lg" type="password" name="fun_senha" v-model="user.fun_senha" >
-                <div class="invalid-feedback color-danger" id="fun_senha" >{{error.fun_senha}}</div>
-            </div>
-            <div class="d-grid gap-2">
-                <button class="btn btn-dark btn-lg" type="submit">Entrar</button>
-            </div>
-            <div class="lg-12 text-center mt-4"><a class="text-dark" href="">Estou com problemas para acessar</a></div>
-        </form>
-    </div>
-</div>
+    <v-container class="px-0 py-0">
+        <v-row no-gutters class="d-flex align-center justify-center h-screen">
+            <v-col
+                cols="12"
+                lg="4"
+                class="d-flex align-center justify-center"
+            >
+                <v-card
+                    flat
+                    theme="light"
+                    density="default"
+                    variant="elevated"
+                    class="mt-12 mt-sm-0 pa-4"
+                >
 
-<div class="loading w-100 h-100" v-show="loading == true">
-    <div class="spinner-border text-white" role="status">
-        <span class="sr-only">Loading...</span>
-    </div>
-    <div class="text-white fs-4">Carregando...</div>
-</div>
+                <h3>Ubarber Places</h3>
+                <h1>Bem vindo ao Ubarber! 👋🏻</h1>
+                <p class="text-disabled">
+                    Para continuar, realize o login em sua conta
+                </p>
+                <v-form :disabled="loading" @submit.prevent="login()" class="mt-10">
+                    <v-col cols="12" class="px-0 py-0">
+                        <v-text-field label="E-mail" variant="outlined" type="email" hide-details="auto" error-messages=""></v-text-field>
+                    </v-col>
+                    
+                    <v-col cols="12" class="px-0 py-0 mt-4">
+                        <v-text-field 
+                            label="Senha" 
+                            variant="outlined" 
+                            :append-inner-icon="visible ? 'mdi-eye-off' : 'mdi-eye'"
+                            :type="visible ? 'text' : 'password'"
+                            @click:append-inner="visible = !visible"
+                            hide-details="auto"
+                        ></v-text-field>
+                    </v-col>
+
+                    <v-col cols="12" class="px-0 py-0 mt-1">
+                        <a href="" class="text-caption text-decoration-none text-black">
+                            Esqueceu sua senha?
+                        </a>
+                    </v-col>
+
+                    <v-col cols="12" class="px-0 py-0 mt-5">
+                        <v-btn color="black" size="large" block>Entrar</v-btn>
+                    </v-col>
+
+                    <v-col cols="12" class="px-0 py-0 mt-5">
+                        <p class="text-center">Não tem cadastro? <a href="" class="text-black font-weight-bold">Cadastre seu Salão</a></p>
+                    </v-col>
+                </v-form>
+
+                </v-card>
+            </v-col>
+        </v-row>
+    </v-container>
 </template>
 
-<style src="@/assets/css/login.css" scoped></style>
+<script lang='ts'>
+import { defineComponent } from 'vue';
+import HelloWorld from '@/components/HelloWorld.vue'
+import axios from 'axios'
 
-<script lang="ts">
-import Login from '@/models/Login'
-export default Login
+const App = defineComponent({
+  components: {
+    HelloWorld
+  },
+
+  data() {
+    return {
+        visible: false,
+        response: false,
+        error: false,
+        consultaExecutada: false,
+        serverUrl: import.meta.env.VITE_SERVER_URL,
+        loading: false,
+    };
+  },
+  methods: {
+    async consultar(){
+      let url = this.serverUrl+'api/test'
+      this.loading = true
+      await axios.get(url)
+      .then((response) => {
+        this.resetMessages()
+        this.consultaExecutada = true
+        if(response.data.response === true){
+          this.response = true
+        }else{
+          this.response = false
+        }
+      })
+      .catch(error => {
+        this.resetMessages()
+        this.consultaExecutada = true
+        this.error = true
+        console.error(error);
+      });
+    },
+
+    login(){
+        alert('Logged!')
+    },
+
+    resetMessages(){
+      this.response = false
+      this.error = false
+      this.consultaExecutada = false
+      this.loading = false
+    }
+  },
+
+  mounted(){
+  }
+});
+
+export default App
 </script>
